+++ examples/stack_viewer_xr.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import nd2
 from qtpy import QtWidgets
 
-from pymmcore_widgets._stack_viewer2._stack_viewer import StackViewer
+from pymmcore_widgets._stack_viewer._stack_viewer import StackViewer
 
 data = nd2.imread("/Users/talley/Downloads/6D_test.nd2", xarray=True, dask=True)
 qapp = QtWidgets.QApplication([])
