+++ src/pymmcore_widgets/_stack_viewer/_backends/_pygfx.py
