+++ src/pymmcore_widgets/_stack_viewer/_dims_slider.py
