+++ src/pymmcore_widgets/_stack_viewer/_protocols.py
