+++ src/pymmcore_widgets/_stack_viewer_v1/_datastore.py
