+++ src/pymmcore_widgets/_stack_viewer_v1/_labeled_slider.py
