+++ src/pymmcore_widgets/experimental.py
@@ -1,3 +1,4 @@
-from ._stack_viewer import StackViewer
+from ._stack_viewer import MDAViewer
+from ._stack_viewer_v1 import StackViewer
 
-__all__ = ["StackViewer"]
+__all__ = ["StackViewer", "MDAViewer"]
