+++ tests/test_stack_viewer1.py
@@ -10,7 +10,8 @@ from useq import MDAEvent, MDASequence
 from vispy.app.canvas import MouseEvent
 from vispy.scene.events import SceneMouseEvent
 
-from pymmcore_widgets._stack_viewer import CMAPS
+from pymmcore_widgets._stack_viewer_v1 import CMAPS
+from pymmcore_widgets._stack_viewer_v1._datastore import QOMEZarrDatastore
 from pymmcore_widgets.experimental import StackViewer
 
 if TYPE_CHECKING:
@@ -177,3 +178,25 @@ def test_not_ready(qtbot: QtBot) -> None:
     # TODO: we should do something here that checks if the loop finishes
     canvas.frameReady(MDAEvent())
     mmcore.mda.run(sequence)
+
+
+def test_reception(qtbot: QtBot) -> None:
+    sequence = MDASequence(
+        channels=[{"config": "DAPI", "exposure": 10}],
+        time_plan={"interval": 0.3, "loops": 3},
+    )
+
+    mmcore = CMMCorePlus.instance()
+
+    datastore = QOMEZarrDatastore()
+    mmcore.mda.events.frameReady.connect(datastore.frameReady)
+    mmcore.mda.events.sequenceFinished.connect(datastore.sequenceFinished)
+    mmcore.mda.events.sequenceStarted.connect(datastore.sequenceStarted)
+
+    with qtbot.waitSignal(datastore.frame_ready, timeout=5000):
+        mmcore.run_mda(sequence, block=False)
+    with qtbot.waitSignal(datastore.frame_ready, timeout=5000):
+        pass
+
+    assert datastore.get_frame(MDAEvent(index={"c": 0, "t": 0})).flatten()[0] != 0
+    qtbot.wait(1000)
